requirements.txt
numpy>=1.21,<2.0
pandas>=1.3,<2.0
pygame>=1.9.3,<3.0
PyYAML>=6.0,<7.0
tensorboard>=2.10,<2.14
torch==1.13.1
